finish version on project
