+++ cloud-client/src/main/java/domain/CurrentLogin.java
@@ -1,0 +1,13 @@
+package domain;
+
+public class CurrentLogin {
+
+    static String currentLogin;
+
+    public static void setCurrentLogin(String login){
+        currentLogin = login;
+    }
+    public static String getCurrentLogin(){
+        return currentLogin;
+    }
+}
