--- cloud-client/src/main/java/supportClass/CurrentLogin.java
@@ -1,13 +1,0 @@
-package supportClass;
-
-public class CurrentLogin {
-
-    static String currentLogin;
-
-    public static void setCurrentLogin(String login){
-        currentLogin = login;
-    }
-    public static String getCurrentLogin(){
-        return currentLogin;
-    }
-}
