+++ cloud-client/src/main/resources/view/mainWindow.fxml
@@ -23,12 +23,14 @@
 
       <Button layoutX="14.0" layoutY="536.0" mnemonicParsing="false" onAction="#sendFileToServer" prefHeight="26.0" prefWidth="158.0" text="SEND" AnchorPane.leftAnchor="10.0" />
       <Button layoutX="210.0" layoutY="536.0" mnemonicParsing="false" onAction="#deleteAction" prefHeight="26.0" prefWidth="158.0" text="DELETE" AnchorPane.leftAnchor="210.0" />
-      <Button layoutX="431.0" layoutY="536.0" mnemonicParsing="false" onAction="#downloadFileFromServer" prefHeight="26.0" prefWidth="158.0" text="DOWNLOAD" AnchorPane.rightAnchor="210.60000000000002" />
-      <Button layoutX="632.0" layoutY="536.0" mnemonicParsing="false" onAction="#sendUpdateFileToServer" prefHeight="26.0" prefWidth="158.0" text="UPDATE" />
+      <Button layoutX="431.0" layoutY="536.0" mnemonicParsing="false" onAction="#downloadFileFromServer" prefHeight="26.0" prefWidth="158.0" text="DOWNLOAD" AnchorPane.rightAnchor="210.0" />
+      <Button layoutX="531.0" layoutY="488.0" mnemonicParsing="false" onAction="#sendUpdateFileToServer" prefHeight="26.0" prefWidth="158.0" text="UPDATE" />
       <Button layoutX="678.0" layoutY="14.0" mnemonicParsing="false" onAction="#sendConnectToServer" prefHeight="26.0" prefWidth="112.0" text="Connect" />
+       <Button layoutX="632.0" layoutY="536.0" mnemonicParsing="false" onAction="#sendDeleteFileFromServer" prefHeight="26.0" prefWidth="158.0" text="DELETE" />
 
       <Text layoutX="158.0" layoutY="115.0" strokeType="OUTSIDE" strokeWidth="0.0" text="My File" textAlignment="CENTER" wrappingWidth="112.0" />
       <Text layoutX="550.0" layoutY="115.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Server" textAlignment="CENTER" wrappingWidth="112.0" />
+
