--- cloud-client/storage/2.txt
@@ -1,8 +1,0 @@
-sFASFASDFSAD FASFASDDFA AFASSF Sasfdasfasdfasdfasdfa s fassdf asfa sdfa
-sFASFASDFSAD FASFASDDFA AFASSF Sasfdasfasdfasdfasdfa s fassdf asfa sdfa sFASFASDFSAD FASFASDDFA AFASSF Sasfdasfasdfasdfasdfa s fassdf asfa sdfa sFASFASDFSAD FASFASDDFA AFASSF Sasfdasfasdfasdfasdfa s fassdf asfa sdfa
-sFASFASDFSAD FASFASDDFA AFASSF Sasfdasfasdfasdfasdfa s fassdf asfa sdfa
-sFASFASDFSAD FASFASDDFA AFASSF Sasfdasfasdfasdfasdfa s fassdf asfa sdfa
-sFASFASDFSAD FASFASDDFA AFASSF Sasfdasfasdfasdfasdfa s fassdf asfa sdfa
-sFASFASDFSAD FASFASDDFA AFASSF Sasfdasfasdfasdfasdfa s fassdf asfa sdfa sFASFASDFSAD FASFASDDFA AFASSF Sasfdasfasdfasdfasdfa s fassdf asfa sdfa
-
-sFASFASDFSAD FASFASDDFA AFASSF Sasfdasfasdfasdfasdfa s fassdf asfa sdfa
