--- common/src/main/java/message/DownloadMessage.java
@@ -1,24 +1,0 @@
-package message;
-
-import java.io.Serializable;
-import java.nio.file.Path;
-
-
-public class DownloadMessage implements Serializable {
-
-    private String fileName;
-    private String login;
-
-    public DownloadMessage(String fileName, String login) {
-        this.fileName = fileName;
-        this.login = login;
-    }
-
-    public String getFileName() {
-        return fileName;
-    }
-
-    public String getLogin() {
-        return login;
-    }
-}
